app-py/requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Database
psycopg2-binary
asyncpg
pgsync # For syncing PostgreSQL with Elasticsearch

# Elasticsearch
elasticsearch

# Environment variables
python-dotenv

# Development dependencies
pytest
pytest-asyncio
httpx